+++ src/main/java/org/fogbowcloud/manager/core/intercomponent/xmpp/requesters/RemoteGetUserQuotaRequest.java
@@ -2,8 +2,6 @@ package org.fogbowcloud.manager.core.intercomponent.xmpp.requesters;
 
 import org.apache.log4j.Logger;
 import org.dom4j.Element;
-import org.fogbowcloud.manager.core.exceptions.FogbowManagerException;
-import org.fogbowcloud.manager.core.exceptions.UnavailableProviderException;
 import org.fogbowcloud.manager.core.exceptions.UnexpectedException;
 import org.fogbowcloud.manager.core.intercomponent.xmpp.XmppErrorConditionToExceptionTranslator;
 import org.fogbowcloud.manager.core.intercomponent.xmpp.IqElement;
@@ -31,34 +29,34 @@ public class RemoteGetUserQuotaRequest implements RemoteRequest<Quota> {
 
     @Override
     public Quota send() throws Exception {
-        IQ iq = createIq();
+        IQ iq = marshal(this.provider, this.federationUser, this.resourceType);
         IQ response = (IQ) PacketSenderHolder.getPacketSender().syncSendPacket(iq);
 
         XmppErrorConditionToExceptionTranslator.handleError(response, this.provider);
-        Quota quota = getUserQuotaFromResponse(response);
+        Quota quota = unmarshalUserQuota(response);
         return quota;
     }
 
-    public IQ createIq() {
+    public static IQ marshal(String provider, FederationUser federationUser, ResourceType resourceType) {
         IQ iq = new IQ(IQ.Type.get);
-        iq.setTo(this.provider);
+        iq.setTo(provider);
 
         Element queryElement = iq.getElement().addElement(IqElement.QUERY.toString(),
                 RemoteMethod.REMOTE_GET_USER_QUOTA.toString());
 
         Element memberIdElement = queryElement.addElement(IqElement.MEMBER_ID.toString());
-        memberIdElement.setText(new Gson().toJson(this.provider));
+        memberIdElement.setText(new Gson().toJson(provider));
 
         Element userElement = queryElement.addElement(IqElement.FEDERATION_USER.toString());
-        userElement.setText(new Gson().toJson(this.federationUser));
+        userElement.setText(new Gson().toJson(federationUser));
 
         Element orderTypeElement = queryElement.addElement(IqElement.INSTANCE_TYPE.toString());
-        orderTypeElement.setText(this.resourceType.toString());
+        orderTypeElement.setText(resourceType.toString());
 
         return iq;
     }
 
-    private Quota getUserQuotaFromResponse(IQ response) throws UnexpectedException {
+    private Quota unmarshalUserQuota(IQ response) throws UnexpectedException {
         Element queryElement = response.getElement().element(IqElement.QUERY.toString());
         String quotaStr = queryElement.element(IqElement.USER_QUOTA.toString()).getText();
